requirements.txt
streamlit==1.21.0
opencv-python-headless==4.8.0.74
Pillow==9.5.0
ultralytics==8.0.62
numpy==1.23.5
blinker==1.6.2
